+++ app/src/main/java/com/shid/mosquefinder/domain/usecases/GetSurahInFrench.kt
@@ -5,11 +5,11 @@ import com.shid.mosquefinder.domain.repository.AyahRepository
 import kotlinx.coroutines.flow.Flow
 
 
-typealias GetSurahInFrenchBaseUseCase = BaseUseCase<Int, Flow<List<Verse>>>
+typealias GetSurahInFrenchBaseUseCase = BaseUseCase<Long, Flow<List<Verse>>>
 
 class GetSurahInFrenchUseCase(private val ayahRepository: AyahRepository) :
     GetSurahInFrenchBaseUseCase {
-    override suspend fun invoke(params: Int): Flow<List<Verse>> {
+    override suspend fun invoke(params: Long): Flow<List<Verse>> {
         return ayahRepository.getSurahInFrench(params)
     }
